chore: reorganize directory structure
