+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.application.ApplicationReader;
+import racingcar.application.RaceManager;
 import racingcar.config.AppConfig;
 import racingcar.config.RaceConfig;
