+++ src/main/java/racingcar/domain/car/CarRegistry.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.domain.car;
 
 import java.util.HashSet;
 import java.util.Set;
