+++ src/test/java/racingcar/CarRegistryTest.java
@@ -2,9 +2,9 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.car.Car;
-import racingcar.car.CarFactory;
-import racingcar.car.CarRegistry;
+import racingcar.domain.car.Car;
+import racingcar.domain.car.CarFactory;
+import racingcar.domain.car.CarRegistry;
 
 import java.util.HashSet;
 import java.util.Set;
