+++ src/test/java/racingcar/CarTest.java
@@ -2,7 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.car.Car;
+import racingcar.domain.car.Car;
 
 import static org.junit.jupiter.api.Assertions.*;
